Correct tracking of which newsfragment files to remove
At the end of a run, towncrier tries to remove the newsfragment files
that it consumed. To do this, it needs to know their filenames.

Previously, we would try to reconstruct the newsfragment filenames
based on our internal data structures. This didn't work (see gh-99).

With this commit, it now records the list of files it actually read
newsfragments from, which is considerably more reliable (and also
simpler).

This did require a bit of an internal refactor. Before, find_fragments
read all the files in the newsfragments dir, and then split_fragments
filtered through them to figure out which are actually newsfragments
and decode their basenames. This meant that the code that knew the
actual full filename was different from the code that knew which files
were actually newsfragments. Moving the filename decoding into
find_fragments solves this.

Fixes gh-99.