Raise an error if a newsfragment collision is detected
Currently (see gh-101), if you have

  newsfragments/1.feature.rst
  newsfragments/1.feature.rst~

then towncrier will silently pick one of them essentially at random.
We should fix gh-101 so that it picks the .rst one, not the .rst~ one,
but this would have been a lot easier to detect if towncrier had
pointed out that these collided. So let's check for that.