enabled fallback of name|package values to project.name
This fallback will only work if the pyproject.toml file
is present.
However, it will work for any combination of pyproject.toml
or towncrier.toml configuration files.

Note, that project.name is a required field in pyproject.toml
files. Hence, the minimal requirement for using towncrier
in python project is a no-op!

Signed-off-by: Nick Papior <nickpapior@gmail.com>